- Obey the single responsibility rule by abstracting payment process to a dedicated class, also OMSDK should handle the process but should not have knowledge of the payment logic. - Code to an interface. - Update README.md
