+++ README.md
@@ -11,17 +11,14 @@ composer require faso-dev/orange-money-burkina-sdk v1.alpha
 ```php
 
     use Fasodev\Sdk\OMSDK;
+    use \Fasodev\Sdk\OrangeMoneyAPI;
 
     require_once __DIR__ . '/../vendor/autoload.php';
 
-    $orangeMoney = OMSDK::init("username", "password", "merchantNumber", OMSDK::ENV_DEV)
-        ->setAmount(1000) //Montant de la transaction
-        ->setOTPCode(121212) //Code otp fourni par l'utilisateur
-        ->setClientNumber(76819212) //Le numero de client
-    ;
-    $result = $orangeMoney
-        ->processPayment() //Enclenchement du processus de paiement
-    ;
+    $processor = new OMSDK(new OrangeMoneyAPI("username", "password", "merchantNumber", OMSDK::ENV_DEV));
+    
+    $result = $processor->handlePayment(); //Enclenchement du processus de paiement
+    
     if ($result->status === 200) {
         echo " paiement effectué";
         echo $result->transID;
